agent/requirements.txt
serpapi>=0.1.5
google-search-results>=2.4.2
requests>=2.31.0
beautifulsoup4>=4.12.3
lxml>=5.2.1
sqlalchemy>=2.0.29
python-dotenv>=1.0.1
rich>=13.7.1
click>=8.1.7
jinja2>=3.1.3
resend>=0.7.2
truststore>=0.9.1
